refactor mcp message handler
